refactor: MovieReviewApp을 컴포넌트 폴더 밖으로 분리
